removing rag specific lightspeed-stack client code
